fix: constants file name and removed commented out code
