+++ packages/coil-extension/src/popup/constants.ts
@@ -1,0 +1,7 @@
+export enum ROUTES {
+  streaming = 'streaming',
+  tipping = 'tipping',
+  tippingConfirm = 'tipping/confirm',
+  tippingComplete = 'tipping/complete',
+  settings = 'settings'
+}
