--- packages/coil-extension/src/popup/contants.ts
@@ -1,12 +1,0 @@
-export enum ROUTES {
-  streaming = 'streaming',
-  // streamingWebMo = 'streaming/monetized',
-  // streamingNoWebMo = 'streaming/notmonetized',
-  // streamingNoMembership = 'streaming/nomembership',
-  // streamingCoil = 'streaming/coil',
-  // streamingCoilDiscover = 'streaming/coil/discover',
-  tipping = 'tipping',
-  tippingConfirm = 'tipping/confirm',
-  tippingComplete = 'tipping/complete',
-  settings = 'settings'
-}
