+++ packages/coil-extension/src/popup/context/routerContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useState } from 'react'
 
-import { ROUTES } from '../contants'
+import { ROUTES } from '../constants'
 
 //
 // Models
